pyproject.toml
[project]
name = "crypto-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ccxt>=4.5.20",
    "fastmcp>=2.13.1",
    "python-dotenv>=1.2.1",
]
